Rename AndroidCommons to AndroidTemplate
