+++ README.md
@@ -1,2 +1,2 @@
-# AndroidCommons
+# AndroidTemplate
 Android の 開発に必要な共通実装です。
