+++ app/build.gradle
@@ -5,7 +5,7 @@ android {
     buildToolsVersion "23.0.2"
 
     defaultConfig {
-        applicationId "jp.hideakisago.androidcommons"
+        applicationId "jp.hideakisago.androidtemplate"
         minSdkVersion 17
         targetSdkVersion 23
         versionCode 1
