+++ app/src/androidTest/java/jp/hideakisago/androidcommons/ApplicationTest.java
@@ -1,4 +1,4 @@
-package jp.hideakisago.androidcommons;
+package jp.hideakisago.androidtemplate;
 
 import android.app.Application;
 import android.test.ApplicationTestCase;
