+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">Android commons</string>
+    <string name="app_name">Android template</string>
 </resources>
